// src/app/views/ActiveCallsView/ActiveCalls.view.interface.ts
export interface ActiveCallsViewProps {
  showContactDisplayPlaceholder?: boolean;
  showRingoutCallControl?: boolean;
  showSwitchCall?: boolean;
  showOtherDevice?: boolean;
  showSmallCallControl?: boolean;
  showTransferCall?: boolean;
  showMergeCall?: boolean;
  showHoldOnOtherDevice?: boolean;
  useV2?: boolean;
  useCallControl?: boolean;
  composeTextRoute?: string;
  callCtrlRoute?: string;
  onCreateContact?: (options: OnCreateContactOptions) => void | Promise<void>;
  onLogCall?: (options: {
    call: HistoryCall | ActiveCall;
    contact: Entity;
    redirect?: boolean;
  }) => void;
  isLoggedContact?: (...args: any) => boolean;
  onCallsEmpty?: () => void;
  onViewContact?: (options: { contact: RouteParams }) => void;
  showViewContact?: boolean;
  getAvatarUrl?: (...args: any) => string | Promise<string | null>;
  sourceIcons?: Record<string, ComponentType | null>;
}

// TODO: move to new ActiveCallsPanel component
export interface ActiveCallsPanelProps {
  currentLocale: string;
  className?: string;
  activeRingCalls: Call[];
  activeOnHoldCalls: Call[];
  activeCurrentCalls: Call[];
  otherDeviceCalls: Call[];
  areaCode: string;
  countryCode: string;
  brand: string;
  showContactDisplayPlaceholder: boolean;
  showSmallCallControl?: boolean;
  formatPhone: (phoneNumber: string) => void;
  onClickToSms?: (
    contact: ToNumber | { name: string; phoneNumber: string },
    isDummyContact?: boolean,
  ) => void;
  onCreateContact?: (options: OnCreateContactOptions) => void | Promise<void>;
  outboundSmsPermission: boolean;
  internalSmsPermission: boolean;
  isLoggedContact?: (...args: any) => boolean;
  onLogCall: (options: {
    call: HistoryCall | ActiveCall;
    contact: Entity;
    redirect?: boolean;
  }) => void;
  showMergeCall?: boolean;
  onMergeCall?: (
    webphoneSessionId: string,
    telephonySessionId: string,
  ) => Promise<void>;
  webphoneAnswer: (
    sessionId: string,
    telephonySessionId: string,
    isHoldAndAnswer?: boolean,
  ) => Promise<void>;
  webphoneReject: (sessionId: string) => Promise<void>;
  webphoneHangup: (
    sessionId: string,
    telephonySessionId: string,
  ) => Promise<void> | void;
  webphoneResume: (
    sessionId: string,
    telephonySessionId: string,
  ) => Promise<void> | void;
  webphoneToVoicemail: (
    sessionId: string,
    telephonySessionId: string,
  ) => Promise<void>;
  webphoneSwitchCall: (
    activeCall: SwitchCallActiveCallParams | ActiveSession,
  ) => Promise<WebphoneSession | void> | undefined;
  webphoneIgnore: (telephonySessionId: string) => Promise<void> | void;
  modalConfirm?: (options: ConfirmModalOptions) => string;
  modalClose?: (id: string) => boolean;
  autoLog: boolean;
  onViewContact: (options: { contact: RouteParams }) => void;
  enableContactFallback?: boolean;
  loggingMap?: any;
  onCallsEmpty: () => void;
  sourceIcons?: Record<string, ComponentType>;
  phoneTypeRenderer?: () => void;
  phoneSourceNameRenderer?: () => void;
  isWebRTC: boolean;
  showSpinner: boolean;
  isSessionAConferenceCall: (sessionId: string) => boolean;
  onCallItemClick: (call: Call) => void;
  getAvatarUrl?: (...args: any) => string | Promise<string | null>;
  conferenceCallParties: (Party & PartyState)[] | null;
  webphoneHold: (
    sessionId: string,
    telephonySessionId: string,
  ) => Promise<any> | void;
  useV2: boolean;
  updateSessionMatchedContact: Webphone['updateSessionMatchedContact'];
  isOnHold: (Session: NormalizedSession) => boolean;
  // CallLog related
  currentLog?: { call: Call; logName: string };
  renderEditLogSection?: () => void;
  renderSaveLogButton?: () => void;
  renderExtraButton?: () => void;
  onSaveCallLog?: () => void;
  onUpdateCallLog?: () => void;
  onCloseLogSection?: () => void;
  // - Notification
  logNotification?: any;
  onCloseNotification?: () => void;
  onDiscardNotification?: () => void;
  onSaveNotification?: () => void;
  onExpandNotification?: () => void;
  showNotiLogButton?: boolean;
  notificationContainerStyles?: string;
  // Contact
  showAvatar?: boolean;
  renderContactName?: (call: Call) => string;
  showOtherDevice?: boolean;
  ringoutHangup: (telephonySessionId: string) => Promise<void> | void;
  ringoutTransfer: (sessionId: string) => void;
  ringoutReject: (sessionId: string) => Promise<void> | void;
  disableLinks: boolean;
  showRingoutCallControl: boolean;
  showMultipleMatch?: boolean;
  showSwitchCall: boolean;
  showTransferCall: boolean;
  showHoldOnOtherDevice: boolean;
  onLogBasicInfoClick?: () => void;
  renderSmallCallContrl?: () => void;
  // customization
  showCallDetail?: boolean;
  showIgnoreBtn?: boolean;
  showHoldAnswerBtn?: boolean;
  useCallDetailV2?: boolean;
  newCallIcon?: boolean;
  useCallControl: boolean;
  clickSwitchTrack: () => void;
  onSwitchCall: (call: Call) => void;
  isWide: boolean;
  allCalls: Call[];
}

export interface ActiveCallsViewOptions {
  component?: typeof ActiveCallsPanel;
  isWide?: boolean;
}